requirements.txt
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
Django==2.2
django-jet==1.0.8
django-modeltranslation==0.16.1
django-rosetta==0.9.5
idna==2.10
jet-bridge-base==0.8.0
Pillow==8.0.1
polib==1.1.0
prompt-toolkit==2.0.9
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
wcwidth==0.2.5
